Update readme in old jobguard location Add exception for building in before-commit.sh

+++ development/tools/before-commit.sh
@@ -40,6 +40,9 @@ fi
 while IFS= read -r -d '' directory
 do
     cmdName=$(basename "${directory}")
+    if [ -a "${cmdName}/nobuild.lock" ]; then
+      continue
+    fi
     ${buildEnv} go build -o "${cmdName}" "${directory}"
     buildResult=$?
     rm "${cmdName}"
